[CLIENT] Restructure Components Directory
Group similar components in one of the following directories:

 * common
 * fields
 * forms
 * providers
