+++ client/components/common/Footer.tsx
