+++ client/components/fields/ComboBoxField.tsx
@@ -1,14 +1,14 @@
 import React from "react";
 import { useField } from "formik";
-import { Combobox, PropsT } from "baseui/ComboBox";
+import { Combobox, PropsT } from "baseui/combobox";
 import { FormControl } from "baseui/form-control";
-import { PartialBy } from "../utils/typeUtil";
+import { PartialBy } from "../../utils/typeUtil";
 
-interface ComboboxProps extends PartialBy<PropsT, "value"> {
+interface ComboboxFieldProps extends PartialBy<PropsT, "value"> {
   label: string;
 }
 
-export const ComboboxField: React.FC<ComboboxProps> = (props) => {
+export const ComboboxField: React.FC<ComboboxFieldProps> = (props) => {
   const [field, { error }, { setValue }] = useField<string>(props.name!);
   return (
     <FormControl label={() => props.label} error={!!error ? error : null}>
