+++ client/components/fields/TextAreaField.tsx
@@ -1,13 +1,13 @@
+import React from "react";
 import { FormControl } from "baseui/form-control";
 import { Textarea, TextareaProps } from "baseui/textarea";
 import { useField } from "formik";
-import React from "react";
 
-interface TextAreaProps extends TextareaProps {
+interface TextAreaFieldProps extends TextareaProps {
   label: string;
 }
 
-export const TextArea: React.FC<TextAreaProps> = (props) => {
+export const TextAreaField: React.FC<TextAreaFieldProps> = (props) => {
   const [field, { error }] = useField<string>(props.name!);
   return (
     <FormControl label={() => props.label} error={!!error ? error : null}>
