+++ client/components/forms/ReleaseRequestForm.tsx
@@ -4,22 +4,24 @@ import { Button } from "baseui/button";
 import { FlexGrid, FlexGridItem } from "baseui/flex-grid";
 import { Heading, HeadingLevel } from "baseui/heading";
 import { Form, Formik, FormikConfig } from "formik";
+
 import {
   EmployeesProfiles,
-  ResourceRequestInput,
+  ReleaseRequestInput,
   useManagersNamesQuery,
-} from "../graphql/types";
-import { CheckBoxStr } from "./CheckBoxStr";
-import { ComboboxField } from "./ComboBoxField";
-import { DatePickerStr } from "./DatePickerStr";
-import { InputField } from "./InputField";
-import { Loading } from "./Loading";
+} from "../../graphql/types";
+import { Loading } from "../common/Loading";
+import { CheckBoxStrField } from "../fields/CheckBoxStrField";
+import { ComboboxField } from "../fields/ComboBoxField";
+import { DatePickerStrField } from "../fields/DatePickerStrField";
+import { InputField } from "../fields/InputField";
+import { TextAreaField } from "../fields/TextAreaField";
 
-interface ResourceRequestProps extends FormikConfig<ResourceRequestInput> {
+interface ReleaseRequestProps extends FormikConfig<ReleaseRequestInput> {
   action: string;
 }
 
-export const ResourceRequestForm: React.FC<ResourceRequestProps> = ({
+export const ReleaseRequestForm: React.FC<ReleaseRequestProps> = ({
   action,
   ...props
 }) => {
@@ -43,7 +45,7 @@ export const ResourceRequestForm: React.FC<ResourceRequestProps> = ({
                 color: theme.colors.primary,
               }}
             >
-              Resource Request
+              Release Request
             </Heading>
           </HeadingLevel>
 
@@ -60,56 +62,49 @@ export const ResourceRequestForm: React.FC<ResourceRequestProps> = ({
                 mapOptionToString={(option: EmployeesProfiles) => option.name}
               />
 
+              <InputField name="employeeName" label="Employee Name" required />
+              <InputField name="employeeId" label="Employee ID" required />
+              <DatePickerStrField
+                name="releaseDate"
+                label="Release Date"
+                required
+              ></DatePickerStrField>
+            </FlexGridItem>
+
+            <FlexGridItem display="flex" flexDirection="column">
+              <InputField
+                name="employeeTitle"
+                label="Employee Title"
+                required
+              />
               <InputField name="function" label="Function" required />
-              <InputField name="title" label="Title" required />
-              <DatePickerStr label="Start Date" name="startDate" required />
-              <DatePickerStr label="End Date" name="endDate" required />
+              <TextAreaField
+                name="releaseReason"
+                label="Release Reason"
+                rows={6}
+                required
+              />
             </FlexGridItem>
 
             <FlexGridItem display="flex" flexDirection="column">
               <InputField
                 name="propability"
                 label="Propability"
-                required
                 type="number"
-              />
-
-              <InputField
-                name="percentage"
-                label="Percentage"
-                required
-                type="number"
-              />
-              <InputField name="status" label="Status" required />
-              <InputField
-                name="assignedResource"
-                label="Assigned Resource"
                 required
               />
               <InputField
-                name="actualPercentage"
-                label="Actual Percentage"
+                name="releasePercentage"
+                label="Release Percentage"
+                type="number"
                 required
-                type="number"
-              />
-            </FlexGridItem>
-
-            <FlexGridItem display="flex" flexDirection="column">
-              <InputField name="replacementFor" label="Replacement For" />
-              <InputField
-                name="requestsCount"
-                label="Requests Count"
-                type="number"
               />
               <InputField
-                name="relatedOpportunity"
-                label="Related Opportunity"
+                name="requestStatus"
+                label="Request Status"
+                required
               />
-              <InputField name="comments" label="Comments" />
-              <br />
-              <CheckBoxStr label="Core Team Member" name="coreTeamMember" />
-              <CheckBoxStr label="Replacenement" name="replacenement" />
-
+              <CheckBoxStrField label="Leaving" name="leaving" />
               <Button
                 type="submit"
                 isLoading={isSubmitting}
