+++ client/pages/create/releaseRequest.tsx
@@ -1,8 +1,8 @@
 import React from "react";
-
 import { useRouter } from "next/dist/client/router";
-import { MainLayout } from "../../components/MainLayout";
-import { ReleaseRequestForm } from "../../components/ReleaseRequestForm";
+
+import { MainLayout } from "../../components/common/MainLayout";
+import { ReleaseRequestForm } from "../../components/forms/ReleaseRequestForm";
 import {
   ReleaseRequestInput,
   useCreateReleaseRequestMutation,
