+++ client/pages/edit/skill/[id].tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { useRouter } from "next/dist/client/router";
 
-import { Loading } from "../../../components/Loading";
-import { MainLayout } from "../../../components/MainLayout";
-import { SkillForm } from "../../../components/SkillForm";
+import { Loading } from "../../../components/common/Loading";
+import { MainLayout } from "../../../components/common/MainLayout";
+import { SkillForm } from "../../../components/forms/SkillForm";
 import {
   UserRole,
   useSkillQuery,
