+++ client/pages/view/certificateProviders.tsx
@@ -12,8 +12,8 @@ import { Cell, Grid } from "baseui/layout-grid";
 import { useRouter } from "next/dist/client/router";
 import { CSVLink } from "react-csv";
 
-import { Loading } from "../../components/Loading";
-import { MainLayout } from "../../components/MainLayout";
+import { Loading } from "../../components/common/Loading";
+import { MainLayout } from "../../components/common/MainLayout";
 import {
   useCertificationsProvidersQuery,
   useDeleteCertificateProviderMutation,
